file_mate/requirements.txt
Pillow
python-magic
click
reportlab
python-magic-bin
setuptools
pypdf